+++ src/commands/blame/disableBlame.ts
@@ -1,0 +1,28 @@
+"use strict";
+
+import { Uri, window } from "vscode";
+import { Command } from "../command";
+import { blameStateManager } from "../../blame/blameStateManager";
+
+/**
+ * Command shown when blame is enabled (eye icon)
+ * Clicking disables blame for the file
+ */
+export class DisableBlame extends Command {
+  constructor() {
+    super("svn.blame.disableBlame");
+  }
+
+  async execute(uri?: Uri): Promise<void> {
+    if (!uri) {
+      const editor = window.activeTextEditor;
+      if (!editor) {
+        return;
+      }
+      uri = editor.document.uri;
+    }
+
+    // Disable blame
+    blameStateManager.setBlameEnabled(uri, false);
+  }
+}